Extract the optional Overview cells to a separate component.
